[ANDROSDK-2078] Refactor toDB functions out of the companion
