+++ core/src/main/java/org/hisp/dhis/android/persistence/attribute/AttributeDB.kt
@@ -133,54 +133,52 @@ internal data class AttributeDB(
             optionAttribute(optionAttribute)
         }.build()
     }
+}
 
-    companion object {
-        fun Attribute.toDB(): AttributeDB {
-            return AttributeDB(
-                uid = uid(),
-                code = code(),
-                name = name(),
-                displayName = displayName(),
-                created = created().dateFormat(),
-                lastUpdated = lastUpdated().dateFormat(),
-                shortName = shortName(),
-                displayShortName = displayShortName(),
-                description = description(),
-                displayDescription = displayDescription(),
-                valueType = valueType()?.name,
-                uniqueProperty = unique(),
-                mandatory = mandatory(),
-                indicatorAttribute = indicatorAttribute(),
-                indicatorGroupAttribute = indicatorGroupAttribute(),
-                userGroupAttribute = userGroupAttribute(),
-                dataElementAttribute = dataElementAttribute(),
-                constantAttribute = constantAttribute(),
-                categoryOptionAttribute = categoryOptionAttribute(),
-                optionSetAttribute = optionSetAttribute(),
-                sqlViewAttribute = sqlViewAttribute(),
-                legendSetAttribute = legendSetAttribute(),
-                trackedEntityAttributeAttribute = trackedEntityAttributeAttribute(),
-                organisationUnitAttribute = organisationUnitAttribute(),
-                dataSetAttribute = dataSetAttribute(),
-                documentAttribute = documentAttribute(),
-                validationRuleGroupAttribute = validationRuleGroupAttribute(),
-                dataElementGroupAttribute = dataElementGroupAttribute(),
-                sectionAttribute = sectionAttribute(),
-                trackedEntityTypeAttribute = trackedEntityTypeAttribute(),
-                userAttribute = userAttribute(),
-                categoryOptionGroupAttribute = categoryOptionGroupAttribute(),
-                programStageAttribute = programStageAttribute(),
-                programAttribute = programAttribute(),
-                categoryAttribute = categoryAttribute(),
-                categoryOptionComboAttribute = categoryOptionComboAttribute(),
-                categoryOptionGroupSetAttribute = categoryOptionGroupSetAttribute(),
-                validationRuleAttribute = validationRuleAttribute(),
-                programIndicatorAttribute = programIndicatorAttribute(),
-                organisationUnitGroupAttribute = organisationUnitGroupAttribute(),
-                dataElementGroupSetAttribute = dataElementGroupSetAttribute(),
-                organisationUnitGroupSetAttribute = organisationUnitGroupSetAttribute(),
-                optionAttribute = optionAttribute(),
-            )
-        }
-    }
+internal fun Attribute.toDB(): AttributeDB {
+    return AttributeDB(
+        uid = uid(),
+        code = code(),
+        name = name(),
+        displayName = displayName(),
+        created = created().dateFormat(),
+        lastUpdated = lastUpdated().dateFormat(),
+        shortName = shortName(),
+        displayShortName = displayShortName(),
+        description = description(),
+        displayDescription = displayDescription(),
+        valueType = valueType()?.name,
+        uniqueProperty = unique(),
+        mandatory = mandatory(),
+        indicatorAttribute = indicatorAttribute(),
+        indicatorGroupAttribute = indicatorGroupAttribute(),
+        userGroupAttribute = userGroupAttribute(),
+        dataElementAttribute = dataElementAttribute(),
+        constantAttribute = constantAttribute(),
+        categoryOptionAttribute = categoryOptionAttribute(),
+        optionSetAttribute = optionSetAttribute(),
+        sqlViewAttribute = sqlViewAttribute(),
+        legendSetAttribute = legendSetAttribute(),
+        trackedEntityAttributeAttribute = trackedEntityAttributeAttribute(),
+        organisationUnitAttribute = organisationUnitAttribute(),
+        dataSetAttribute = dataSetAttribute(),
+        documentAttribute = documentAttribute(),
+        validationRuleGroupAttribute = validationRuleGroupAttribute(),
+        dataElementGroupAttribute = dataElementGroupAttribute(),
+        sectionAttribute = sectionAttribute(),
+        trackedEntityTypeAttribute = trackedEntityTypeAttribute(),
+        userAttribute = userAttribute(),
+        categoryOptionGroupAttribute = categoryOptionGroupAttribute(),
+        programStageAttribute = programStageAttribute(),
+        programAttribute = programAttribute(),
+        categoryAttribute = categoryAttribute(),
+        categoryOptionComboAttribute = categoryOptionComboAttribute(),
+        categoryOptionGroupSetAttribute = categoryOptionGroupSetAttribute(),
+        validationRuleAttribute = validationRuleAttribute(),
+        programIndicatorAttribute = programIndicatorAttribute(),
+        organisationUnitGroupAttribute = organisationUnitGroupAttribute(),
+        dataElementGroupSetAttribute = dataElementGroupSetAttribute(),
+        organisationUnitGroupSetAttribute = organisationUnitGroupSetAttribute(),
+        optionAttribute = optionAttribute(),
+    )
 }
