+++ core/src/main/java/org/hisp/dhis/android/persistence/dataapproval/DataApprovalDB.kt
@@ -65,16 +65,14 @@ internal data class DataApprovalDB(
             state?.let { state(DataApprovalState.valueOf(it)) }
         }.build()
     }
+}
 
-    companion object {
-        fun DataApproval.toDB(): DataApprovalDB {
-            return DataApprovalDB(
-                workflow = workflow()!!,
-                organisationUnit = organisationUnit()!!,
-                period = period()!!,
-                attributeOptionCombo = attributeOptionCombo()!!,
-                state = state()?.name,
-            )
-        }
-    }
+internal fun DataApproval.toDB(): DataApprovalDB {
+    return DataApprovalDB(
+        workflow = workflow()!!,
+        organisationUnit = organisationUnit()!!,
+        period = period()!!,
+        attributeOptionCombo = attributeOptionCombo()!!,
+        state = state()?.name,
+    )
 }
